Don't take screenshots of http error pages
phantomjs does not consider a response to be unsuccessful unless it didn't get a
valid http response at all (network interrution, timeout, etc). We have to hook
the callbacks to even figure out what the response code was. This represents
a fork from upstream wraith as I doubt they'll want to include this unless we
make it configurable and off by default.